adds landing page component and login links to nav
