+++ src/components/css/header-bar.css
@@ -3,92 +3,102 @@
 @import url('https://fonts.googleapis.com/css?family=Bitter:700');
 
 .content-header-reduced {
-	margin-left: 300px; 
-	width:50%;
-	padding: 0px 10px;
-background: #faf7f4;
-
+  margin-left: 300px;
+  width: 50%;
+  padding: 0px 10px;
+  background: #faf7f4;
 }
 
 .content-header-expanded {
-	margin-left: 10px; 
-	padding: 0px 10px;
+  margin-left: 10px;
+  padding: 0px 10px;
 }
 
 .content-header-expanded a {
-text-decoration: none;
+  text-decoration: none;
 }
 
-.content-header-reduced  a{
-	text-decoration: none;
+.content-header-reduced a {
+  text-decoration: none;
 }
 
-
-.period{
-	color:purple;
-	font-size: 50px;
-	padding: 0px 4px 0px 4px;
+.period {
+  color: purple;
+  font-size: 50px;
+  padding: 0px 4px 0px 4px;
 }
-
-
 
 .header-title {
-	letter-spacing: -5px;
-	font-family: 'Bitter', serif;
-	color: #000000;
-	font-size: 56px;
-	text-align: center;
-	font-weight: 100;
-	margin: 0 auto;
-	text-transform: lowercase;
-	padding-left: 80px;
-
+  letter-spacing: -5px;
+  font-family: 'Bitter', serif;
+  color: #000000;
+  font-size: 56px;
+  text-align: center;
+  font-weight: 100;
+  margin: 0 auto;
+  text-transform: lowercase;
+  padding-left: 80px;
 }
 
-
-
-.content-header-black{
-	background-color: #000000;
+.content-header-black {
+  background-color: #000000;
+  padding: 10px;
 }
 
-.content-header-black a{
-	text-decoration: none;
+.content-header-black a {
+  text-decoration: none;
 }
 
 .header-title-black {
-
-	font-family: 'Bitter', serif;
-	color: white;
-	font-size: 60px;;
-	text-align: center;
-	font-weight: 100;
-	margin: 0 auto;
-	text-transform: lowercase;
+  font-family: 'Bitter', serif;
+  color: white;
+  font-size: 50px;
+  text-align: left;
+  font-weight: 100;
+  margin: 0 auto;
+  text-transform: lowercase;
+  width: 50%;
+  padding-left: 1%;
+  box-sizing: border-box;
 }
 
+.header-title-black,
+.content-header-black nav,
+.content-header-black ul {
+  display: inline-block;
+  vertical-align: middle;
+}
+
+.content-header-black nav {
+  margin-right: -200px;
+  width: 50%;
+  text-align: right;
+  padding-right: 25px;
+  box-sizing: border-box;
+}
+
+.content-header-black ul li {
+  display: inline;
+  vertical-align: middle;
+  margin-right: 18px;
+}
 
 @media only screen and (min-width: 315px) {
-	.content-header-reduced {
-		margin-left: 100px; 
-		width:100%;
-		padding: 0px 10px;
-	
-	
-	}
+  .content-header-reduced {
+    margin-left: 100px;
+    width: 100%;
+    padding: 0px 10px;
+  }
 
-
-	.header-title {
-		letter-spacing: -5px;
-		font-family: 'Bitter', serif;
-		color: #000000;
-		font-size: 56px;
-		text-align: center;
-		font-weight: 100;
-		margin: 0 auto;
-		text-transform: lowercase;
-		padding-left: 10px;
-	
-	}
-
-
+  .header-title {
+    letter-spacing: -5px;
+    font-family: 'Bitter', serif;
+    color: #000000;
+    font-size: 56px;
+    text-align: center;
+    font-weight: 100;
+    margin: 0 auto;
+    text-transform: lowercase;
+    padding-left: 10px;
+  }
 }
